YD-627 Refactored the user service to split in 4
The user service grew to a very large service (nearly 1000 lines). To make it more manageable, it's now split into 4 parts, related to retrieval, addition, deletion and update. The dependencies are as follows:

* UserDeletionService: not used by the others
* UserAdditionService: not used by the others
* UserUpdateService: used by UserAdditionService
* UserRetrievalService: used by UserAdditionService and UserUpdateService
